pci: Declare structures and IOMMU operation for PRI
The API is deliberately designed to be similar to ATS.
We define a struct that stores a function pointer to the device's callback.
Registering and unregistering a notifier is done using a pair of functions
that will be added in a future commit of this series.

An IOMMU can support PRI by implementing the iommu_pri_request_page
operation declared in IOMMUMemoryRegionClass.

Signed-off-by: Clement Mathieu--Drif <clement.mathieu--drif@eviden.com>